requirements.txt
# Основные зависимости
beautifulsoup4==4.12.3
lxml==5.3.0
click==8.1.7
colorama==0.4.6
pyyaml==6.0.2
pydantic==2.10.0

# Для разработки
pytest==8.3.0
black==24.8.0
pylint==3.2.0
mypy==1.11.0